Datagrid now has an ErrorBox component in which it shows errors related to the delete button (admin cannot delete another admin and server is down)

+++ src/components/Datagrid.jsx
@@ -2,10 +2,12 @@ import * as React from 'react';
 import { DataGrid } from '@mui/x-data-grid';
 import Button from '@mui/material/Button';
 import { useState, useEffect } from 'react';
+import ErrorBox from "./ErrorBox";
 
 export default function DataGridUsers(props) {
   
   const [rows, setRows] = useState([]);
+  const [deleteButtonError, setDeleteError] = useState('');
   useEffect(() => {
     setRows(props.rows)
   }, [])
@@ -49,12 +51,17 @@ export default function DataGridUsers(props) {
           <Button variant="contained" color="secondary" size="small" style={{marginLeft: 16}}
                   aria-label={'deleteUser' + params.row.id}
                   onClick={() => {
+                    if (params.row.user_type === 'admin') {
+                      setDeleteError("You cannot delete an admin");
+                      return;
+                    }
+                    setDeleteError('');
                     props.apiHandler.deleteUser(params.row.id)
                     .then((res) => { 
                       setRows(rows.filter((user) => user.id !== params.row.id))
                     })
                     .catch((error) => {
-                      console.log(error)
+                      setDeleteError("Server is not available. Try again later.");
                     })
                   }}>
             Delete User
@@ -66,6 +73,7 @@ export default function DataGridUsers(props) {
 
   return (
     <div >
+      <ErrorBox errorString={deleteButtonError}/>
       <DataGrid
         rows={rows}
         columns={columns}
@@ -75,7 +83,6 @@ export default function DataGridUsers(props) {
         disableSelectionOnClick
         style={{ height: "700px", widht: "100%"}}
         columnBuffer={7}
-
       />
     </div>
   );
